requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pynput>=1.7.6
pyaudio>=0.2.13
webrtcvad>=2.0.10
SpeechRecognition>=3.10.0
youtube-transcript-api>=1.2.0
google-api-python-client>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
feedparser>=6.0.0
elevenlabs>=2.16.0
pydub>=0.25.1
anthropic>=0.18.0
simpleaudio>=1.0.4
